+++ src/nestjs/src/modules/database/game/gameInfo/dto.ts
@@ -1,5 +1,5 @@
 import { IsString, IsNotEmpty, IsArray, IsOptional } from "class-validator";
-import { PlayerScoreEntity } from "../player-score/entity";
+import { PlayerScoreEntity } from "../playerScore/entity";
 
 export class DBGameInfoPost {
 	@IsNotEmpty()
