+++ src/nestjs/src/modules/database/game/playerScore/entity.ts
